Fix excessively clever syntax in logo-xhr.svg
While it seems technically correct and worked in Firefox and Chrome,
it did not work in Inkscape or rsvg.  Also, it's *really* obscure.

Inkscape (silently) throws a fit and gives up on getting any more
valid segments out of the path data specification, which turns out to
be the correct approach to invalid paths[1] (though it should inform
the user).

rsvg just draws a funny arc and keeps reading the path.

I've reported https://bugs.debian.org/740784 (rsvg) and
https://bugs.debian.org/740785 (Inkscape) about the bad rendering.

[1] http://www.w3.org/TR/SVG/implnote.html#PathElementImplementationNotes