Invoke checker with CHECKER_PARAMS env variable
The CHECKER_PARAMS environment variable is empty by default. The value
can be set per-repo in .travis.yml files.

The only purpose for it at this point is to allow the checker
`filterpattern` param to be specified so that particular error/warning
messages to be filtered out (dropped/ignored/suppressed)—for example,
the “Text run is not in Unicode Normalization Form C.” warnings that
are causing noise for the Encoding spec build.