Ensure black bullet/disc for dfn-popup references
This change ensures that when a dfn is inside a bulleted list (as for example
most every dfn in the HTML spec Dependencies section) the references displayed
in its dfn popup are shown with the normal/usual filled-black-bullet list-item
markers (top-level "disc" list-item markers in CSS terms).

Without this change, the references for a dfn that’s inside a bulleted list
end up shown with unfilled bullets (second-level "circle" markers).